requirements.txt
sentence-transformers
langchain-community
langchain
chromadb
PyPDF2
langchain_mistralai
tf-keras
langchain-huggingface
streamlit
faiss-cpu
tensorflow-cpu
torch